Day_1

+++ Basic/que_3.py
@@ -2,7 +2,7 @@
 
 print("Enter Rectangle Length and Breadth for Area calculation")
 
-length = int(input("The Length of Rectangle is:"))
-breadth = int(input("The Bredth of Rectangle is:"))
+length = float(input("The Length of Rectangle is:"))
+breadth = float(input("The Bredth of Rectangle is:"))
 area = length * breadth 
 print(f"The area of Rectangle having length {length} and breadth {breadth} is {area}")
